move client to separate file
