--- util.go
@@ -1,18 +1,0 @@
-package web3
-
-import "math/big"
-
-var (
-	weiPerGO   = big.NewInt(1e18)
-	weiPerGwei = big.NewInt(1e9)
-)
-
-// WeiAsBase converts w wei in to the base unit, and formats it as a decimal fraction with full precision (up to 18 decimals).
-func WeiAsBase(w *big.Int) string {
-	return new(big.Rat).SetFrac(w, weiPerGO).FloatString(18)
-}
-
-// WeiAsGwei converts w wei in to gwei, and formats it as a decimal fraction with full precision (up to 9 decimals).
-func WeiAsGwei(w *big.Int) string {
-	return new(big.Rat).SetFrac(w, weiPerGwei).FloatString(9)
-}
